tb3289-teimurazi-bakuradze-winter-2024-001-se-181-course-project-master/src/main/java/banking/WithdrawProcessor.java
package banking;

public class WithdrawProcessor extends CommandProcessor {
    private double withdrawAmount;
    private int accountId;

    WithdrawProcessor(Bank bank) {
        super(bank);
    }

    public void processWithdraw(String[] commandParts) {
        parseWithdrawCommand(commandParts);
        bank.getAccount(accountId).withdraw(withdrawAmount);
    }

    public void parseWithdrawCommand(String[] commandParts) {
        this.accountId = Integer.parseInt(commandParts[1]);
        this.withdrawAmount = Double.valueOf(commandParts[2]);

    }
}
